pyproject.toml
[build-system]
requires = ["setuptools>=48", "setuptools_scm[toml]>=6.3.1"]
build-backend = "setuptools.build_meta"

[project]
name = "document_to_podcast"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dynamic = ["version"]
dependencies = [
  "beautifulsoup4",
  "huggingface-hub",
  "llama-cpp-python",
  "loguru",
  "PyPDF2[crypto]",
  "python-docx",
  "streamlit",
]

[project.optional-dependencies]
docs = [
  "mkdocs",
  "mkdocs-material",
  "mkdocstrings-python",
]

tests = [
  "pytest>=8,<9",
  "pytest-sugar>=0.9.6",
]

[project.urls]
Documentation = "https://mozilla-ai.github.io/OpenNotebookLLM/"
Issues = "https://github.com/mozilla-ai/OpenNotebookLLM/issues"
Source = "https://github.com/mozilla-ai/OpenNotebookLLM"

[tool.setuptools.packages.find]
exclude = ["tests", "tests.*"]
where = ["src"]
namespaces = false

[tool.setuptools_scm]
